refactor: split utils.go into ws.go, http.go, and url.go for SRP
